integrate pftq video extension f1 model
